+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/session/signout/SignOutTask.kt
@@ -16,25 +16,64 @@
 
 package im.vector.matrix.android.internal.session.signout
 
+import android.content.Context
 import im.vector.matrix.android.api.auth.data.Credentials
 import im.vector.matrix.android.internal.SessionManager
 import im.vector.matrix.android.internal.auth.SessionParamsStore
+import im.vector.matrix.android.internal.crypto.CryptoModule
+import im.vector.matrix.android.internal.database.RealmKeysUtils
+import im.vector.matrix.android.internal.di.CryptoDatabase
+import im.vector.matrix.android.internal.di.SessionDatabase
+import im.vector.matrix.android.internal.di.UserCacheDirectory
+import im.vector.matrix.android.internal.di.UserMd5
 import im.vector.matrix.android.internal.network.executeRequest
+import im.vector.matrix.android.internal.session.SessionModule
+import im.vector.matrix.android.internal.session.cache.ClearCacheTask
 import im.vector.matrix.android.internal.task.Task
+import im.vector.matrix.android.internal.worker.WorkManagerUtil
+import timber.log.Timber
+import java.io.File
 import javax.inject.Inject
 
 internal interface SignOutTask : Task<Unit, Unit>
 
-internal class DefaultSignOutTask @Inject constructor(private val credentials: Credentials,
+internal class DefaultSignOutTask @Inject constructor(private val context: Context,
+                                                      private val credentials: Credentials,
                                                       private val signOutAPI: SignOutAPI,
                                                       private val sessionManager: SessionManager,
-                                                      private val sessionParamsStore: SessionParamsStore) : SignOutTask {
+                                                      private val sessionParamsStore: SessionParamsStore,
+                                                      @SessionDatabase private val clearSessionDataTask: ClearCacheTask,
+                                                      @CryptoDatabase private val clearCryptoDataTask: ClearCacheTask,
+                                                      @UserCacheDirectory private val userFile: File,
+                                                      private val realmKeysUtils: RealmKeysUtils,
+                                                      @UserMd5 private val userMd5: String) : SignOutTask {
 
     override suspend fun execute(params: Unit) {
+        Timber.d("SignOut: send request...")
         executeRequest<Unit> {
             apiCall = signOutAPI.signOut()
         }
-        sessionParamsStore.delete(credentials.userId)
+
+        Timber.d("SignOut: release session...")
         sessionManager.releaseSession(credentials.userId)
+
+        Timber.d("SignOut: cancel pending works...")
+        WorkManagerUtil.cancelAllWorks(context)
+
+        Timber.d("SignOut: delete session params...")
+        sessionParamsStore.delete(credentials.userId)
+
+        Timber.d("SignOut: clear session data...")
+        clearSessionDataTask.execute(Unit)
+
+        Timber.d("SignOut: clear crypto data...")
+        clearCryptoDataTask.execute(Unit)
+
+        Timber.d("SignOut: clear file system")
+        userFile.deleteRecursively()
+
+        Timber.d("SignOut: clear the database keys")
+        realmKeysUtils.clear(SessionModule.DB_ALIAS_PREFIX + userMd5)
+        realmKeysUtils.clear(CryptoModule.DB_ALIAS_PREFIX + userMd5)
     }
 }