+++ vector/src/main/java/im/vector/riotx/core/di/ScreenComponent.kt
@@ -21,7 +21,6 @@ import androidx.lifecycle.ViewModelProvider
 import dagger.BindsInstance
 import dagger.Component
 import im.vector.fragments.keysbackup.restore.KeysBackupRestoreFromPassphraseFragment
-import im.vector.matrix.android.api.session.Session
 import im.vector.riotx.core.preference.UserAvatarPreference
 import im.vector.riotx.features.MainActivity
 import im.vector.riotx.features.crypto.keysbackup.restore.KeysBackupRestoreFromKeyFragment
@@ -42,15 +41,14 @@ import im.vector.riotx.features.home.createdirect.CreateDirectRoomKnownUsersFrag
 import im.vector.riotx.features.home.group.GroupListFragment
 import im.vector.riotx.features.home.room.detail.RoomDetailFragment
 import im.vector.riotx.features.home.room.detail.readreceipts.DisplayReadReceiptsBottomSheet
-import im.vector.riotx.features.home.room.detail.timeline.action.MessageActionsBottomSheet
-import im.vector.riotx.features.home.room.detail.timeline.action.MessageMenuFragment
-import im.vector.riotx.features.home.room.detail.timeline.action.QuickReactionFragment
-import im.vector.riotx.features.home.room.detail.timeline.action.ViewEditHistoryBottomSheet
-import im.vector.riotx.features.home.room.detail.timeline.action.ViewReactionBottomSheet
+import im.vector.riotx.features.home.room.detail.timeline.action.*
 import im.vector.riotx.features.home.room.filtered.FilteredRoomsActivity
 import im.vector.riotx.features.home.room.list.RoomListFragment
 import im.vector.riotx.features.invite.VectorInviteView
+import im.vector.riotx.features.link.LinkHandlerActivity
 import im.vector.riotx.features.login.LoginActivity
+import im.vector.riotx.features.login.LoginFragment
+import im.vector.riotx.features.login.LoginSsoFallbackFragment
 import im.vector.riotx.features.media.ImageMediaViewerActivity
 import im.vector.riotx.features.media.VideoMediaViewerActivity
 import im.vector.riotx.features.navigation.Navigator
@@ -65,20 +63,14 @@ import im.vector.riotx.features.roomdirectory.createroom.CreateRoomActivity
 import im.vector.riotx.features.roomdirectory.createroom.CreateRoomFragment
 import im.vector.riotx.features.roomdirectory.picker.RoomDirectoryPickerFragment
 import im.vector.riotx.features.roomdirectory.roompreview.RoomPreviewNoPreviewFragment
-import im.vector.riotx.features.settings.VectorSettingsActivity
-import im.vector.riotx.features.settings.VectorSettingsAdvancedNotificationPreferenceFragment
-import im.vector.riotx.features.settings.VectorSettingsHelpAboutFragment
-import im.vector.riotx.features.settings.VectorSettingsNotificationPreferenceFragment
-import im.vector.riotx.features.settings.VectorSettingsNotificationsTroubleshootFragment
-import im.vector.riotx.features.settings.VectorSettingsPreferencesFragment
-import im.vector.riotx.features.settings.VectorSettingsSecurityPrivacyFragment
+import im.vector.riotx.features.settings.*
 import im.vector.riotx.features.settings.push.PushGatewaysFragment
 
 @Component(dependencies = [VectorComponent::class], modules = [AssistedInjectModule::class, ViewModelModule::class, HomeModule::class])
 @ScreenScope
 interface ScreenComponent {
 
-    fun session(): Session
+    fun activeSessionHolder(): ActiveSessionHolder
 
     fun viewModelFactory(): ViewModelProvider.Factory
 
@@ -134,6 +126,10 @@ interface ScreenComponent {
 
     fun inject(publicRoomsFragment: PublicRoomsFragment)
 
+    fun inject(loginFragment: LoginFragment)
+
+    fun inject(loginSsoFallbackFragment: LoginSsoFallbackFragment)
+
     fun inject(sasVerificationIncomingFragment: SASVerificationIncomingFragment)
 
     fun inject(quickReactionFragment: QuickReactionFragment)
@@ -141,6 +137,8 @@ interface ScreenComponent {
     fun inject(emojiReactionPickerActivity: EmojiReactionPickerActivity)
 
     fun inject(loginActivity: LoginActivity)
+
+    fun inject(linkHandlerActivity: LinkHandlerActivity)
 
     fun inject(mainActivity: MainActivity)