+++ CHANGES.md
@@ -2,7 +2,8 @@ Changes in RiotX 0.5.0 (2019-XX-XX)
 ===================================================
 
 Features:
- -
+ - Handle M_CONSENT_NOT_GIVEN error (#64)
+ - Auto configure homeserver and identity server URLs of LoginActivity with a magic link
 
 Improvements:
  - Reduce default release build log level, and lab option to enable more logs.
@@ -14,6 +15,7 @@ Bugfix:
  - Fix crash due to missing informationData (#535)
  - Progress in initial sync dialog is decreasing for a step and should not (#532)
  - Fix rendering issue of accepted third party invitation event
+ - All current notifications were dismissed by mistake when the app is launched from the launcher
 
 Translations:
  -